requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# NLP and AI
langchain>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Visualization (optional)
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
